bill: 修复表单编辑问题；system: 去掉小红书二维码；habit: 打卡单位支持小数

+++ db/habit.go
@@ -6,6 +6,6 @@ type Habit struct {
 	ID        uint64    `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time `json:"created_at" gorm:"index"`
 	Name      string    `json:"name"`   // 名称
-	Number    int       `json:"number"` // 数量
+	Number    float64   `json:"number"` // 数量
 	Unit      string    `json:"unit"`   // 单位
 }
